long rework 2

+++ src/main/java/ru/netology/javaqa/javaqamvn/services/Main.java
@@ -8,7 +8,7 @@ public class Main {
         long averageSales = service.averageSum(sales);
         System.out.println("Средняя сумма продаж: " + averageSales);
 
-        long monthsBelowAverage = service.belowAverage(sales, averageSales);
+        int monthsBelowAverage = service.belowAverage(sales, averageSales);
         System.out.println("Количество месяцев с продажами ниже среднего: " + monthsBelowAverage);
